Revert "Setup UI structure (#177)"
This reverts commit cd4a46b8922ec5cddc8c3d63a8896c6f419a8d20.
