+++ web/package.json
@@ -9,19 +9,12 @@
     "start:prod": "NODE_ENV=production node build/server.js"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.17",
-    "@emotion/styled": "^10.0.17",
     "d3-fetch": "^1.1.2",
-    "emotion-theming": "^10.0.19",
     "express": "^4.17.1",
-    "moment": "^2.24.0",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-map-gl": "^5.1.3",
-    "react-router-dom": "^5.0.1",
-    "react-semantic-ui-datepickers": "^2.0.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^0.88.1"
+    "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.0",
@@ -33,7 +26,6 @@
     "@types/webpack-env": "^1.13.9",
     "prettier": "^1.18.2",
     "razzle": "^3.0.0",
-    "razzle-plugin-scss": "^3.0.0",
     "razzle-plugin-typescript": "^3.0.0",
     "ts-jest": "^24.0.2",
     "tslint": "^5.18.0",
