+++ web/src/App.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 
-import App from 'components/App'
+import App from './App'
 
 import { MemoryRouter } from 'react-router-dom'
