--- web/src/components/Home/index.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import './index.css'
-
-const Home = () => (
-  <div className='Home'>
-    <div className='Home-header'>
-      <h2>Welcome to WOAQ</h2>
-    </div>
-  </div>
-)
-
-export default Home
