+++ web/src/server.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { renderToString } from 'react-dom/server'
 import { StaticRouter } from 'react-router-dom'
 
-import App from 'components/App'
+import App from './App'
 
 let assets: any
 
@@ -28,7 +28,7 @@ const server = express()
     <head>
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta charSet='utf-8' />
-        <title>West Oakland Air Quality Project</title>
+        <title>Razzle TypeScript</title>
         <meta name="viewport" content="width=device-width, initial-scale=1">
         ${
           assets.client.css
