+++ web/tsconfig.json
@@ -2,25 +2,19 @@
   "compilerOptions": {
     "allowJs": true,
     "allowUnreachableCode": false,
+    "noFallthroughCasesInSwitch": true,
     "allowSyntheticDefaultImports": true,
-    "baseUrl": "src",
     "downlevelIteration": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "jsx": "react",
     "lib": ["es6", "es2015", "es2017", "dom"],
     "module": "esnext",
-    "moduleResolution": "Node",
-    "noFallthroughCasesInSwitch": true,
+    "moduleResolution": "node",
     "noImplicitAny": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
     "noUnusedLocals": false,
-    "paths": {
-      "components/*": ["components/*"],
-      "routes": ["routes"],
-      "theme": ["theme"],
-    },
     "sourceMap": true,
     "strictNullChecks": true,
     "strict": true,
